Reorganize code for tests
